[security][CVE-2022-27809] Builtins should always take int64_t, not int
Builtins are allowed to take 32 or 64-bit integers
(int/int32_t versus int64_t). It's not clear why we allow this, as
integers from Hack are always 64-bit. A builtin taking an int will
just have the upper 32-bits truncated. Depending on the value, this
might change a negative number to a positive one, defeating various
input validity checks.
Remove support for 32-bit integers. Change all of the builtins which
took or returned an int to use int64_t instead. This required a few
fixups here and there, mainly around format strings. If the builtin
uses 32-bits ints internally, I left them. At least the truncation is
now obvious.

Test Plan:
New unit test. Beforehand this large negative input would
be turned into a large positive number, bypassing the "input cannot be
negative" check. Now that it takes int64_t, it is correctly recognized
as a negative number.

---
Initializes m_arr in ArrayInitBase

ArrayInitBase's protected constructor does not initialize
m_arr (presumably because a derived class is meant to do so). However
if a derived class' constructor throws, we'll run the ArrayInitBase
destructor, which will try to release m_arr. If m_arr hasn't been
initialized yet, it can contain garbage and we'll try to release
random memory. The protected constructor should initialize m_arr to
nullptr.

Test Plan:
New unit test which previously failed with ASAN builds

+++ hphp/runtime/ext/datetime/ext_datetime.h
@@ -224,17 +224,17 @@ bool HHVM_FUNCTION(date_default_timezone_set,
                    const String& name);
 Variant HHVM_FUNCTION(timezone_name_from_abbr,
                       const String& abbr,
-                      int gmtoffset = -1,
-                      int isdst = 1);
+                      int64_t gmtoffset = -1,
+                      int64_t isdst = 1);
 String HHVM_FUNCTION(timezone_version_get);
 
 ///////////////////////////////////////////////////////////////////////////////
 // datetime
 
 bool HHVM_FUNCTION(checkdate,
-                   int month,
-                   int day,
-                   int year);
+                   int64_t month,
+                   int64_t day,
+                   int64_t year);
 Variant HHVM_FUNCTION(date_create,
                       const Variant& time = uninit_variant,
                       const Variant& timezone = uninit_variant);
@@ -254,4 +254,3 @@ Array HHVM_FUNCTION(date_sun_info,
 
 ///////////////////////////////////////////////////////////////////////////////
 }
-
